requirements.txt
beautifulsoup4==4.12.3
blinker==1.4
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
cliapp==1.20180812.1
cmdtest==0.32+git
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
et-xmlfile==1.1.0
gyp==0.1
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.3.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.0.0
oauthlib==3.2.0
openpyxl==3.1.5
pandas==2.2.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
systemd-python==234
ttystatus==0.38
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
wadllib==1.3.6
zipp==1.0.0
